+++ libs/react/src/__mocks__/handlers.ts
@@ -19,8 +19,10 @@ import { getNewsDefault } from './newsMock';
 import { getActionMock, getActionsMock } from './action';
 import { v4 as uuidv4 } from 'uuid';
 import {
+  getAllComplianceMocks,
   getComplianceCountsMock,
   getComplianceMock,
+  getComplianceMockByName,
   getOrganizationComplianceMock,
   getOrganizationComplianceMockByName,
   getQuestionnaireContainerMock,
@@ -332,5 +334,14 @@ export const handlers = [
 
   rest.get(getApiUrl('/compliance/:name/organizations/:orgId/responses/counts'), (req, res, ctx) => {
     return res(ctx.json(getComplianceCountsMock()));
+  }),
+
+  rest.get(getApiUrl('/compliance/all/organizations/:orgId'), (req, res, ctx) => {
+    return res(ctx.json(getAllComplianceMocks()));
+  }),
+
+  rest.get(getApiUrl('/compliance/:name'), (req, res, ctx) => {
+    const name = req.params.name as string;
+    return res(ctx.json(getComplianceMockByName(name)));
   }),
 ];