+++ libs/react/src/__mocks__/storyMockProvider.tsx
@@ -9,7 +9,7 @@ import { ComplianceManagerStatus, ErrorType } from '@coldpbc/enums';
 import { WizardContext, WizardContextType } from '@coldpbc/components';
 import ColdMQTTContext from '../context/coldMQTTContext';
 import { mockMQTTContext } from './mqtt/mockMQTTContext';
-import { defaultMqttDataHandler, defaultMqttTopics, getSectionGroupList } from './mqtt';
+import { defaultMqttDataHandler, defaultMqttTopics } from './mqtt';
 import {
   ColdComplianceManagerContext,
   ColdComplianceQuestionnaireContext,
@@ -18,8 +18,8 @@ import {
   ComplianceQuestionnaireContextType,
 } from '@coldpbc/context';
 import { getAllFilesMock } from './filesMock';
-import { getComplianceCountsMock, getComplianceMock, getQuestionnaireSidebarComplianceMock } from './complianceMock';
-import { ComplianceManagerCountsPayload, QuestionnaireQuestionComplianceResponse } from '@coldpbc/interfaces';
+import { getComplianceMock, getComplianceCountsMock, getQuestionnaireSidebarComplianceMock } from './complianceMock';
+import { AIDetails, ComplianceManagerCountsPayload, ComplianceSidebarPayload, QuestionnaireQuestionComplianceResponse } from '@coldpbc/interfaces';
 
 export interface StoryMockProviderProps {
   handlers?: RestHandler<MockedRequest<DefaultBodyType>>[];
@@ -71,9 +71,6 @@ export const StoryMockProvider = (props: PropsWithChildren<StoryMockProviderProp
   const [showOverviewModal, setShowOverviewModal] = React.useState<boolean>(props.complianceManagerContext?.showOverviewModal || false);
 
   const complianceManagerContextData: ComplianceManagerData = {
-    mqttComplianceSet: getSectionGroupList({
-      name: 'rei_pia_2024',
-    }),
     name: 'rei_pia_2024',
     files: {
       data: getAllFilesMock(),
@@ -84,7 +81,6 @@ export const StoryMockProvider = (props: PropsWithChildren<StoryMockProviderProp
       mutate: () => Promise.resolve(),
     } as SWRResponse<any[], any, any>,
     currentAIStatus: undefined,
-    orgCompliances: undefined,
     complianceCounts: {
       data: getComplianceCountsMock(),
       error: undefined,
@@ -93,6 +89,15 @@ export const StoryMockProvider = (props: PropsWithChildren<StoryMockProviderProp
       isLoading: false,
       mutate: () => Promise.resolve(),
     } as SWRResponse<ComplianceManagerCountsPayload, any, any>,
+    sectionGroups: {
+      data: getQuestionnaireSidebarComplianceMock(),
+      error: undefined,
+      revalidate: () => {},
+      isValidating: false,
+      isLoading: false,
+      mutate: () => Promise.resolve(),
+    } as SWRResponse<ComplianceSidebarPayload, any, any>,
+    compliance: getComplianceMock().find(c => c.name === 'rei_pia_2024'),
     ...props.complianceManagerContext?.data,
   };
 
@@ -107,14 +112,7 @@ export const StoryMockProvider = (props: PropsWithChildren<StoryMockProviderProp
 
   const [complianceQuestionnaireFocusQuestion, setComplianceQuestionnaireFocusQuestion] = React.useState<{
     key: string;
-    aiDetails: {
-      ai_response: QuestionnaireQuestionComplianceResponse['ai_response'];
-      ai_answered?: boolean;
-      ai_attempted?: boolean;
-      value?: any;
-      questionAnswerSaved: boolean;
-      questionAnswerChanged: boolean;
-    };
+    aiDetails: AIDetails;
   } | null>(props.complianceQuestionnaireContext?.focusQuestion ?? null);
 
   const [complianceQuestionnaireScrollToQuestion, setComplianceQuestionnaireScrollToQuestion] = React.useState<string | null>(
@@ -131,7 +129,6 @@ export const StoryMockProvider = (props: PropsWithChildren<StoryMockProviderProp
       isLoading: false,
       mutate: () => Promise.resolve(),
     } as SWRResponse<any, any, any>,
-    complianceDefinition: getComplianceMock().find(compliance => compliance.name === 'rei_pia_2024'),
     ...props.complianceQuestionnaireContext,
     scrollToQuestion: complianceQuestionnaireScrollToQuestion,
     setScrollToQuestion: props.complianceQuestionnaireContext?.setScrollToQuestion ?? setComplianceQuestionnaireScrollToQuestion,